+++ src/components/Nav.tsx
@@ -4,6 +4,9 @@ import {
   FaSmile,
   FaLightbulb,
   FaSmileBeam,
+  FaCog,
+  FaCloudSun,
+  FaCloudSunRain,
 } from 'react-icons/fa';
 import { Link, useLocation } from '@reach/router';
 import globalStyles from '@styles/index';
@@ -11,11 +14,12 @@ import routes from '../utils/routes';
 
 function Item({ route, children }: { route: string; children: ReactNode }) {
   const { pathname } = useLocation();
+  const isActive = pathname === route;
 
   return (
     <li
       className={`    
-        tracking-widest
+        tracking-wide
         h-full w-full
       `}
     >
@@ -25,11 +29,24 @@ function Item({ route, children }: { route: string; children: ReactNode }) {
           ${globalStyles.transitions}
           block h-full
           flex flex-col lg:flex-row justify-center items-center space-y-1 lg:space-x-2
-          text-sm font-medium
+          text-sm
           text-onPrimary
+          transform
           hover:no-underline
+          ${
+            isActive
+              ? `
+          text-light
+           scale-105
+          -translate-y-px
+          font-semibold
+          `
+              : `
+          scale-100
+          translate-y-px
+          `
+          }
         `}
-        style={pathname === route ? { color: 'var(--color-light)' } : {}}
       >
         {children}
       </Link>
@@ -43,6 +60,7 @@ export default function Nav() {
     text-2xl
     md:text-xl
     block
+    ${globalStyles.transitions}
   `;
 
   const Name = ({ name }: { name: string }) => (
@@ -57,8 +75,16 @@ export default function Nav() {
   );
 
   const navItems = [
-    <Item route={routes.home} key="nav-blog">
+    <Item route={routes.home} key="nav-home">
       {pathname === routes.home ? (
+        <FaCloudSunRain className={iconStyles} />
+      ) : (
+        <FaCloudSun className={iconStyles} />
+      )}
+      <Name name="Home" />
+    </Item>,
+    <Item route={routes.blog} key="nav-blog">
+      {pathname === routes.blog ? (
         <FaLightbulb className={iconStyles} />
       ) : (
         <FaRegLightbulb className={iconStyles} />
@@ -83,16 +109,28 @@ export default function Nav() {
 
       <Name name="About" />
     </Item>,
+    <Item route={routes.uses} key="nav-uses">
+      <FaCog
+        className={`
+      ${iconStyles} text-2xl transform ${globalStyles.transitions}
+      ${pathname === routes.uses ? ' rotate-180' : 'rotate-0'}
+      `}
+      />
+
+      <Name name="Uses" />
+    </Item>,
   ];
 
   function getUnderlineOffset() {
     let off = navItems.length;
     if (pathname === routes.home) {
       off = 0;
-    } else if (pathname === routes.about) {
+    } else if (pathname === routes.blog) {
       off = 1;
-    } else if (pathname === routes.contact) {
+    } else if (pathname === routes.about) {
       off = 2;
+    } else if (pathname === routes.uses) {
+      off = 3;
     }
 
     return off === navItems.length
@@ -105,10 +143,8 @@ export default function Nav() {
   return (
     <nav
       className={`
-      w-full h-full lg:max-w-sm
+      w-full h-full lg:max-w-screen-md
       flex flex-col
-      
-
       `}
     >
       <ul
@@ -119,18 +155,16 @@ export default function Nav() {
         `}
       >
         {navItems}
-
-        {/* <Item route={routes.contact}>
-          <FaFire className={iconStyles} />
-          <Name name="Contact" />
-        </Item> */}
       </ul>
       <hr
         style={{ marginLeft: underlineOffset }}
         className={`
         ${globalStyles.transitions} 
         border-0
-        mt-auto h-1 w-1/2
+        mt-auto h-1 
+        
+        
+        w-1/4
         ${underlineOffset ? 'bg-onPrimary' : 'bg-primary'}
         `}
       />
