+++ src/pages/about/index.mdx
@@ -1,0 +1,40 @@
+---
+title: About
+description: About Sean Keever.
+---
+
+# About
+
+![A picture that Sean Keever reuses over and over.](./me.jpg)
+
+My name is **Sean Keever**, builder of Skies! I am a developer currently living
+in Seattle, WA. I graduated from the University of Washington with a
+**bachelor's degree in Computer Engineering**.
+
+> Fun fact. I worked on a farm to help pay for college. I wrote about it in my
+> [personal statement](/blog/person-statement) for admission to UW.
+
+![Standing in front of the the Paul G. Allen school of computer science and engineering](graduation.jpg)
+
+I'm **currently working at OfferUp** from home. I work with the Discovery team
+who is responsible for search within the app.
+
+In terms of development, I am interested in
+
+- Frontend web development
+- Microservice architectures
+- Open source software
+
+In terms of my other interests:
+
+- **Music**. I play guitar and piano. I used to be big into recording
+  electronic, hip hop, and pop music. These days I play casually!
+- **Video games**. I currently have a Switch. Probably my favorite game is
+  Legend of Zelda for the memories it brings. I am excited about Microsoft
+  Flight Simulator, Gran Turismo 7, and Last of Us.
+- **Learning things**. I love learning new things from books, podcasts, and
+  videos. When I feel somewhat confident about a subject, I like to
+  [blog about it](/blog).
+
+  If there's anything you want me to write about in my [blog](/blog), let me
+  know on Twitter [@swkeever](https://twitter.com/swkeever).
