+++ src/utils/routes.js
@@ -3,10 +3,11 @@ const withLegal = (slug) => `/legal/${slug}`;
 module.exports = {
   blog: '/blog',
   home: '/',
-  contact: '/contact/',
-  about: '/about/',
-  resume: '/resume/',
+  contact: '/contact',
+  about: '/about',
+  resume: '/resume',
   sitemap: '/sitemap.xml',
+  uses: '/uses',
   rss: '/rss.xml',
   legal: {
     privacyPolicy: withLegal('privacy-policy'),
